Aggiunta pianeti e relative lune fatto, appena avete tempo date uno sguardo

+++ src/Position.java
@@ -1,18 +1,18 @@
 
 public class Position {
-	private float x;
-	private float y;
+	private double x;
+	private double y;
 	
-	public float getX() {
+	public double getX() {
 		return x;
 	}
-	public void setX(float x) {
+	public void setX(double x) {
 		this.x = x;
 	}
-	public float getY() {
+	public double getY() {
 		return y;
 	}
-	public void setY(float y) {
+	public void setY(double y) {
 		this.y = y;
 	}
 }
